+++ mods/include/mods/game-mode/game-mode.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <string>
+#include <symbols/minecraft.h>
 
 int get_seed_from_string(std::string str);
-extern bool gamemode_implement_save_locally_button;
+Screen *game_mode_create_screen(bool creating_new_world = true, const std::string &old_world_name = "");