+++ mods/include/mods/misc/misc.h
@@ -62,3 +62,4 @@ std::string misc_base64_encode(const std::string &data);
 std::string misc_base64_decode(const std::string &input);
 
 static constexpr int line_height = 8;
+static constexpr int world_size = 16; // In Chunks